Move the form widgets of the ui module into the core namespace
1.19 already moved ActiveForm, ActiveField, ContentHiddenCheckbox,
ContentVisibilitySelect and SortOrderField out of the ui module into
humhub\widgets\form. The remaining eight widgets follow them, the
CodeMirror asset bundle goes to humhub\assets and TabbedFormModel to
humhub\interfaces — so core classes like humhub\widgets\form\ActiveField
and humhub\widgets\bootstrap\FormTabs no longer reach back into a module.

Every old name stays as a deprecated subclass until 1.21.

IconValidator is dropped instead: a single commit in its whole history,
no user in the core or in any known module, and a body of three lines.
