+++ protected/humhub/compat/HForm.php
@@ -12,8 +12,8 @@ use humhub\helpers\Html;
 use humhub\modules\content\widgets\richtext\RichText;
 use humhub\modules\content\widgets\richtext\RichTextField;
 use humhub\modules\file\components\FileManager;
-use humhub\modules\ui\form\widgets\DatePicker;
-use humhub\modules\ui\form\widgets\MultiSelect;
+use humhub\widgets\form\DatePicker;
+use humhub\widgets\form\MultiSelect;
 use humhub\widgets\form\ActiveField;
 use humhub\widgets\form\ActiveForm;
 use humhub\widgets\form\SortOrderField;
