+++ protected/humhub/modules/admin/views/logging/index.php
@@ -6,8 +6,8 @@ use humhub\libs\DateHelper;
 use humhub\modules\admin\assets\LogAsset;
 use humhub\modules\admin\models\forms\LogFilterForm;
 use humhub\modules\admin\models\Log;
-use humhub\modules\ui\form\widgets\DatePicker;
-use humhub\modules\ui\form\widgets\MultiSelect;
+use humhub\widgets\form\DatePicker;
+use humhub\widgets\form\MultiSelect;
 use humhub\widgets\form\ActiveForm;
 use yii\data\Pagination;
 use yii\helpers\Url;
