+++ protected/humhub/modules/content/widgets/ContentTagPicker.php
@@ -12,7 +12,7 @@ namespace humhub\modules\content\widgets;
 use humhub\modules\content\components\ContentContainerActiveRecord;
 use humhub\modules\content\components\ContentTagActiveQuery;
 use humhub\modules\content\models\ContentTag;
-use humhub\modules\ui\form\widgets\BasePicker;
+use humhub\widgets\form\BasePicker;
 use Yii;
 use yii\db\ActiveRecord;
