+++ protected/humhub/modules/ui/form/widgets/DurationPicker.php
@@ -2,56 +2,15 @@
 
 /**
  * @link https://www.humhub.org/
- * @copyright Copyright (c) 2018 HumHub GmbH & Co. KG
+ * @copyright Copyright (c) HumHub GmbH & Co. KG
  * @license https://www.humhub.com/licences
  */
 
 namespace humhub\modules\ui\form\widgets;
 
-use DateTime;
-use humhub\assets\JqueryTimeEntryAsset;
-use humhub\helpers\Html;
-use yii\helpers\Json;
-
 /**
- * DurationPicker renders an UI form widget to select time durations.
- *
- * @since 1.3
- * @author Luke
+ * @deprecated since 1.20, use \humhub\widgets\form\DurationPicker instead. Will be removed in 1.21.
  */
-class DurationPicker extends JsInputWidget
+class DurationPicker extends \humhub\widgets\form\DurationPicker
 {
-    /**
-     * @inheritdoc
-     */
-    public function run()
-    {
-        $view = $this->getView();
-        $id = $this->options['id'];
-
-        $options = [
-            'show24Hours' => true,
-            'unlimitedHours' => true,
-            'defaultTime' => '01:00',
-            'timeSteps' => [1, 15],
-            'spinnerImage' => '',
-        ];
-
-        JqueryTimeEntryAsset::register($view);
-        $view->registerJs("$('#{$id}').timeEntry(" . Json::htmlEncode($options) . ");");
-
-        Html::addCssClass($this->options, 'form-control');
-
-        if ($this->model !== null) {
-            return Html::activeTextInput($this->model, $this->attribute, $this->getOptions());
-        } else {
-            return Html::input($this->name, $this->value, $this->getOptions());
-        }
-    }
-
-    public static function getDuration(DateTime $start, DateTime $end)
-    {
-        $duration = $start->diff($end);
-        return $duration->h . ':' . $duration->m;
-    }
 }
