+++ protected/humhub/widgets/form/ActiveField.php
@@ -9,7 +9,6 @@
 namespace humhub\widgets\form;
 
 use humhub\helpers\Html;
-use humhub\modules\ui\form\widgets\JsInputWidget;
 use humhub\modules\ui\icon\widgets\Icon;
 use yii\base\Widget;
 use yii\helpers\ArrayHelper;
